requirements.txt
numpy>=1.25.0,<2.0.0
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-huggingface==0.3.0
faiss-cpu==1.11.0
sentence-transformers==4.1.0
PyPDF2==3.0.1
streamlit==1.32.0
python-dotenv==1.1.0
huggingface-hub==0.33.0
transformers==4.52.4
torch==2.7.1 